+++ NEWS.md
@@ -14,6 +14,7 @@ editor_options:
 -   Change defaults used for `p.adjust.method` so that they depend on the argument passed to `contrasts`.
 -   Accept in `stat_correlation()`, `stat_poly_eq()`, `stat_ma_eq()`, and `stat_multcomp()`, `p.digits = Inf` as a request to use scientific notation for `p.value.label` (suggested by _wbvguo_, Wenbin Guo).
 -   Fix bug in `stat_multcomp()`: **wrongly encoded letters** in Tukey contrasts in case of grouping factors with more than nine levels.
+-   Fix bug in `stat_multcomp()`: **warning** issued by `mvtnorm::pmvnorm()` because of convergence failure in Tukey contrasts in case of grouping factors with more than approximately 5 to 7 levels. Convergence failure also meant slightly different _P_-values returned in different runs in these cases.
 -   Fix bug in `stat_multcomp()`: off-plot letter labels with Tukey contrasts in some plots with more than five groups.
 
 # ggpmisc 0.5.5